gen_image_upload_proj/requirements.txt
supabase==2.16.0
python-dotenv==1.1.1